add types file
