+++ redux/features/counter-slice.ts
@@ -1,11 +1,8 @@
 'use client';
 
+import { CounterState } from '@/app/resources/types';
 import { createSlice } from '@reduxjs/toolkit'
 import type { PayloadAction } from '@reduxjs/toolkit'
-
-export interface CounterState {
-  value: number
-}
 
 const initialState: CounterState = {
   value: 0,
